Extract ESC/P2 dispatch to printer_dispatch.cpp
Moves the giant ProcessCommandChar() switch (~1080 lines) plus the
two bit-image helpers it drives (SetupBitImage, PrintBitGraph) out
to src/hardware/printer/printer_dispatch.cpp.

printer.cpp shrinks from ~2200 lines of mixed concerns to ~770 lines
covering: lifecycle (ctor/dtor/ResetPrinter), codepage selection,
the C-style PRINTER_* IO entry points + PRINTER_Configure /
PRINTER_FormFeed / PRINTER_Reset.

The remaining file split (per the plan): the ESC opcode naming
(step 11) stays deferred. Naming all 192 case labels carries a
high typo risk for very modest readability gain; the existing
'case 0xNN: // Description' comments are already navigable.

After this commit:
  printer.cpp          770   lifecycle + codepage + IO entry points
  printer_dispatch.cpp 1274  ESC/P2 dispatch + bit-image helpers
  printer_glyph.cpp    192   FillPalette + UpdateFont + Blit/DrawLine
  printer_ps.cpp       203   OutputPagePostScript + ASCII85 encoder
  printer_png.cpp      95    write_png_page
  printer_glue.cpp     260   staging integration (config, IO, mapper)
  printer_charmaps.cpp 486   codepage tables (unchanged)
