+++ src/hardware/CMakeLists.txt
@@ -41,6 +41,7 @@ target_sources(dosboxcommon PRIVATE
 
   printer/printer.cpp
   printer/printer_charmaps.cpp
+  printer/printer_dispatch.cpp
   printer/printer_glue.cpp
   printer/printer_glyph.cpp
   printer/printer_png.cpp
